Move PageLayout from layouts to components - use NavLink for Account link (sidebar)
