+++ apps/web/components/PageLayout/AuthLayout/NavLink.tsx
@@ -3,8 +3,8 @@ import NextLink from 'next/link'
 import {useRouter} from 'next/router'
 import type {ComponentType} from 'react'
 
-import {ExternalLink} from '../../../components/ExternalLink'
-import type {SvgIconProps} from '../../../components/icons'
+import {ExternalLink} from '../../ExternalLink'
+import type {SvgIconProps} from '../../icons'
 
 interface NavLinkProps {
   name: string
