+++ apps/web/pages/auth.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image'
 import {RedirectTo} from '../components/common-components'
 import {supabase} from '../contexts/common-contexts'
 import {useSession} from '../contexts/session-context'
-import {PageLayout} from '../layouts/PageLayout'
+import {PageLayout} from '../components/PageLayout'
 import {VeniceTheme} from '../styles/themes'
 
 export default function AuthScreen() {
